+++ src/frontend/src/App.jsx
@@ -7,7 +7,7 @@ import Header from './header/index';
 import useAuthCheck from './hook/useAuthCheck';
 import Routers from './routers';
 import CookieBanner from './cookieBanner/cookie';
-// import './testServer';
+import './testServer';
 import Footer from './footer';
 import './App.scss';
 import './locales/i18n';
