+++ src/frontend/src/TouristPlaces/UnauthorizedPopup.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 export default function UnauthorizedPopup() {
   return (
     <div className="unauthorized">
-      <div className="unauthorized-background" />
+      <div className="tourist-places-background" />
       <div className="unauthorized-content">
         <h2 className="unauthorized-content__title">Увага!</h2>
         <p className="unauthorized-content__text">Карта туристичних місць доступна тільки авторизованим користувачам</p>
