+++ src/frontend/src/hook/useAuthCheck.js
@@ -14,7 +14,7 @@ function useAuthCheck() {
   function updateAuthValue(value) {
     setAuth(value);
     if (!cookies.get('rememberMe')) {
-      cookies.set('rememberMe', value);
+      cookies.set('rememberMe', JSON.stringify(value));
     }
   }
   return { auth, updateAuthValue };
